feat: adiciona integração Teams automática e resolve problemas de áudio
- Adiciona teams_integration.py com monitoramento automático
- Resolve problema as_loopback em audio_recorder.py
- Implementa teams-monitor comando no Raycast
- Adiciona file_manager.py para gerenciamento de arquivos
- Resolve compatibilidade py-cpuinfo com fallback
- Atualiza README.md com status 100% funcional
- Adiciona CLAUDE.md para instruções de desenvolvimento

+++ src/core/settings_manager.py
@@ -38,7 +38,7 @@ class WhisperSettings:
     model_size: str = "base"
     language: Optional[str] = None
     device: str = "auto"
-    compute_type: str = "float16"
+    compute_type: str = "float32"
     beam_size: int = 5
     temperature: float = 0.0
     condition_on_previous_text: bool = True
